+++ source/primitives/PointDelta.h
@@ -4,27 +4,27 @@
 
 class PointDelta : public Point {
 	private:
-		float magnitude;
+		double magnitude;
 	public:
 		PointDelta();
-		PointDelta(float xval, float yval, float magnitude);
-		PointDelta(Point point, float magnitude);
+		PointDelta(double xval, double yval, double magnitude);
+		PointDelta(Point point, double magnitude);
 		PointDelta(const PointDelta& that);
 		~PointDelta();
 		PointDelta& operator=(const PointDelta& that);
-		PointDelta operator/(const float& num);
-		PointDelta operator*(const float& num);
+		PointDelta operator/(const double& num);
+		PointDelta operator*(const double& num);
 		PointDelta onlyX();
 		PointDelta onlyY();
 		PointDelta negate();
 		bool operator>(const int& val);
 		bool operator<(const int& val);
-		float calcAngle(Point point);
-		float getMagnitude();
+		double calcAngle(Point point);
+		double getMagnitude();
 		Point getXPoint();
 		Point getYPoint();
 		void reduceToBounds();
 		void operator-=(Point delta);
 		void operator+=(Point delta);
-		void setMagnitude(float magnitude);
+		void setMagnitude(double magnitude);
 };
