+++ source/primitives/Vector.cpp
@@ -1,8 +1,8 @@
 #include "Vector.h"
 
-Vector::Vector(float angle) : Point(cos(angle), sin(angle)) {}
+Vector::Vector(double angle) : Point(cos(angle), sin(angle)) {}
 
-Vector::Vector(float x, float y) : Point(x, y) {}
+Vector::Vector(double x, double y) : Point(x, y) {}
 
 Vector::~Vector() {}
 
@@ -14,19 +14,19 @@ Vector Vector::operator/(const double& num) {
 	return (*this) * (1.0 / num);
 }
 
-float Vector::dot(const Vector other) {
+double Vector::dot(const Vector other) {
 	return (this->xval * other.xval) + (this->yval * other.yval);
 }
 
-float Vector::operator*(const Vector& other) {
+double Vector::operator*(const Vector& other) {
 	return this->dot(other);
 }
 
-float Vector::magnitude() const {
+double Vector::magnitude() const {
 	return sqrt(pow(this->xval, 2) + pow(this->yval, 2));
 }
 
-float Vector::angleBetween(const Vector other) {
+double Vector::angleBetween(const Vector other) {
 	return acos(this->dot(other) / (this->magnitude() * other.magnitude()));
 }
